# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Casilleros ESCOM",
    "version": "1.0.0",
    "description": "Documentación de la API para la gestión de casilleros en ESCOM."
  },
  "servers": [
    { "url": "http://localhost:3001" }
  ],
  "paths": {
...
    "/documentos": {
      "post": {
        "summary": "Subir documento PDF para una solicitud",
        "tags": ["Documentos"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_solicitud": { "type": "integer", "example": 1 },
                  "tipo": { "type": "string", "enum": ["comprobante horario", "credencial vigente"] },
                  "archivo": { "type": "string", "format": "binary" }
                },
                "required": ["id_solicitud", "tipo", "archivo"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Documento subido correctamente" },
          "400": { "description": "Faltan campos requeridos o archivo" },
          "500": { "description": "Error al subir documento" }
        }
      }
    },
    "/documentos/solicitud/{id_solicitud}": {
      "get": {
        "summary": "Obtener documentos de una solicitud",
        "tags": ["Documentos"],
        "parameters": [
          {
            "name": "id_solicitud",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de documentos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Documento" }
                }
              }
            }
          },
          "500": { "description": "Error al obtener documentos" }
        }
      }
    },
...
    "/casilleros": {
      "get": {
        "summary": "Obtener todos los casilleros",
        "tags": ["Casilleros"],
        "responses": {
          "200": { "description": "Lista de casilleros" },
          "500": { "description": "Error al obtener casilleros" }
        }
      },
      "post": {
        "summary": "Crear casillero (solo coordinador)",
        "tags": ["Casilleros"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numero": { "type": "integer", "example": 101 },
                  "ubicacion": { "type": "string", "example": "Planta baja" }
                },
                "required": ["numero", "ubicacion"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Casillero creado correctamente" },
          "400": { "description": "Faltan campos requeridos" },
          "409": { "description": "El número de casillero ya existe" },
          "500": { "description": "Error al crear casillero" }
        }
      }
    },
    "/casilleros/disponibles": {
      "get": {
        "summary": "Obtener casilleros disponibles",
        "tags": ["Casilleros"],
        "responses": {
          "200": { "description": "Lista de casilleros disponibles" },
          "500": { "description": "Error al obtener casilleros disponibles" }
        }
      }
    },
    "/casilleros/{id_casillero}/disponibilidad": {
      "put": {
        "summary": "Cambiar disponibilidad de casillero",
        "tags": ["Casilleros"],
        "parameters": [
          {
            "name": "id_casillero",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "disponible": { "type": "boolean", "example": true }
                },
                "required": ["disponible"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Disponibilidad actualizada correctamente" },
          "400": { "description": "Valor de disponibilidad inválido" },
          "404": { "description": "Casillero no encontrado" },
          "500": { "description": "Error al actualizar disponibilidad" }
        }
      }
    },
    "/asignaciones": {
      "post": {
        "summary": "Crear asignación de casillero (pago validado y casillero disponible)",
        "tags": ["Asignaciones"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_pago": { "type": "integer", "example": 1 },
                  "id_casillero": { "type": "integer", "example": 10 }
                },
                "required": ["id_pago", "id_casillero"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Asignación creada correctamente" },
          "400": { "description": "El pago no está validado o el casillero no está disponible" },
          "500": { "description": "Error al asignar casillero" }
        }
      },
      "get": {
        "summary": "Obtener todas las asignaciones (coordinador)",
        "tags": ["Asignaciones"],
        "responses": {
          "200": { "description": "Lista de asignaciones" },
          "500": { "description": "Error al obtener asignaciones" }
        }
      }
    },
    "/asignaciones/usuario/{id_usuario}": {
      "get": {
        "summary": "Obtener asignaciones de un usuario",
        "tags": ["Asignaciones"],
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de asignaciones del usuario" },
          "500": { "description": "Error al obtener asignaciones" }
        }
      }
    },
    "/casilleros/{id_casillero}": {
      "delete": {
        "summary": "Eliminar casillero (solo si no está asignado, solo coordinador)",
        "tags": ["Casilleros"],
        "parameters": [
          {
            "name": "id_casillero",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Casillero eliminado correctamente" },
          "400": { "description": "El casillero está asignado y no puede eliminarse" },
          "404": { "description": "Casillero no encontrado" },
          "500": { "description": "Error al eliminar casillero" }
        }
      }
    },
...
    "/asignaciones/{id_asignacion}/pdf": {
      "get": {
        "summary": "Descargar constancia PDF de asignación de casillero",
        "tags": ["Asignaciones"],
        "parameters": [
          {
            "name": "id_asignacion",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID de la asignación de casillero"
          }
        ],
        "responses": {
          "200": {
            "description": "PDF generado correctamente",
            "content": {
              "application/pdf": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "404": { "description": "Asignación no encontrada" },
          "500": { "description": "Error al generar PDF" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}